+++ src/server/maupod/cmd/audioscan/.maupod.yml
@@ -1,0 +1,23 @@
+pgconn: user=postgres host=localhost password=nevermind sslmode=disable
+retries: 10
+delay: 2
+port: ':8000'
+stores:
+  - name: local-hd
+#    file-system
+    type: 0
+#    location: /media/mau/music-library/music
+#    location: /Users/mau/Downloads/music/Sting - Bring On The Night (1986) [2CD]
+    location: /Users/mau/Downloads/music
+  - name: local-hd-image
+    type: 2
+#    location: /media/mau/music-library/album-covers
+    location: /Users/mau/Downloads/artwork
+#  - name: music bucket
+#    s3
+#    type: 1
+#    location: my-music-bucket
+extensions:
+  - .mp3
+  - .flac
+  - .m4a
